refactor(katana): return error as default for state commitment provider (#2967)
